upload files
if you want to use this so 
first you create one assets folder here and make sure folder name same as mine 
second you create a video folder in assets folder and this folder name are also same as mine 
third you download a video from anywhere and place it in the video folder and video name  is "Fake Hacking! Fake Virus! Pretend to be a Pro Hacker!.mp4" is your video is not same the video is not appear on web